LayoutTests/inspector/timeline/timeline-script-tag-1-expected.txt

CONSOLE MESSAGE: line 8: iframe loaded
Tests the Timeline API instrumentation of an HTML script tag.


ParseHTML
ParseHTML
----> EvaluateScript
--------> TimeStamp : SCRIPT TAG
EvaluateScript Properties:
{
    startTime : <number>
    data : {
        url : <string>
        lineNumber : <number>
    }
    children : <object>
    endTime : <number>
    type : "EvaluateScript"
    usedHeapSize : <number>
    totalHeapSize : <number>
    counters : <object>
}
ParseHTML
